Completed collision resolution system. For now, it just separates overlapping entities.

+++ core/Entities/Components/CollisionBox.cs
@@ -1,6 +1,6 @@
 namespace CaravansCore.Entities.Components;
 
 public record CollisionBox(
-    double Width,
-    double Height
+    float Width,
+    float Height
 ) : IComponent;
